Moved value query to its own file
