+++ Thrive.csproj
@@ -22,6 +22,7 @@
   <ItemGroup>
     <Compile Include="src\general\FilterArgument.cs" />
     <Compile Include="src\general\IFilter.cs" />
+    <Compile Include="src\general\ValueQuery.cs" />
     <Compile Include="src\gui_common\dialogs\queries\ISnapshotable.cs" />
     <Compile Include="src\gui_common\dialogs\queries\IValueQuery.cs" />
     <Compile Include="src\gui_common\dialogs\FilterArgumentPopupMenu.cs" />
